requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.21.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.73.0
anyio==4.12.1
appdirs==1.4.4
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.5
blinker==1.9.0
build==1.4.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.1.8
crewai==1.9.3
crewai-tools==1.9.3
cryptography==46.0.5
curl_cffi==0.13.0
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
discord.py==2.7.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
et_xmlfile==2.0.0
fastuuid==0.14.0
filelock==3.24.3
flatbuffers==25.12.19
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.72.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
httpx-sse==0.4.3
huggingface_hub==0.36.2
identify==2.6.16
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
instructor==1.12.0
Jinja2==3.1.6
jiter==0.10.0
json5==0.10.0
json_repair==0.25.3
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
lancedb==0.5.7
litellm==1.59.12
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.23.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
multitasking==0.0.12
narwhals==2.17.0
nodeenv==1.10.0
numpy==2.4.2
oauthlib==3.3.1
onnxruntime==1.24.2
openai==1.83.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pdfminer.six==20251230
pdfplumber==0.11.9
peewee==4.0.0
pillow==12.1.1
platformdirs==4.9.2
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.5.1
propcache==0.4.1
protobuf==5.29.6
py==1.11.0
pyarrow==23.0.1
pybase64==1.4.3
pycparser==3.0
pydantic==2.11.10
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.11.0
pylance==0.9.18
PyMuPDF==1.26.7
pypdfium2==5.5.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-discovery==1.1.0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.22
pytube==15.0.0
pytz==2025.2
PyYAML==6.0.3
ratelimiter==1.2.0.post0
referencing==0.37.0
regex==2024.9.11
requests==2.32.5
requests-oauthlib==2.0.0
retry==0.9.2
rich==14.3.3
rpds-py==0.30.0
semver==3.0.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
sse-starlette==3.3.2
starlette==0.52.1
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.2
tomli_w==1.1.0
tornado==6.5.4
tqdm==4.67.3
typer==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uv==0.9.30
uvicorn==0.41.0
uvloop==0.22.1
virtualenv==21.1.0
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
yarl==1.22.0
yfinance==1.2.0
youtube-transcript-api==1.2.4
zipp==3.23.0
